+++ cs-framework/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Codestar Framework Changelog
 
 ## PRE-RELEASE 1.0.2
+- Added: Taxonomy framework clear form elements after saving
 - Improved: Icons select field for avoid conflict
 
 ## 1.0.1